functions/digest-generator/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.*
google-cloud-firestore==2.*
google-cloud-translate==3.*
google-cloud-storage==2.*
google-generativeai==0.3.*
vertexai==1.*